refactor: `ProcedimentoService`
- Adicionar atributo constante para mensagens de erro
- Definir atributos como constantes
- Corrigir construtor: injetar "mapper"
- Adicionar método "mostrarProcedimento"
- Renomear métodos

+++ src/main/java/br/com/karol/sistema/service/ProcedimentoService.java
@@ -18,15 +18,18 @@ import br.com.karol.sistema.repository.ProcedimentoRepository;
 @Transactional
 public class ProcedimentoService {
 
-    private ProcedimentoRepository repository;
-    private ProcedimentoMapper mapper;
+    private final String NOT_FOUND_MESSAGE = "Procedimento não encontrado";
+
+    private final ProcedimentoRepository repository;
+    private final ProcedimentoMapper mapper;
 
     public ProcedimentoService(ProcedimentoRepository repository, ProcedimentoMapper mapper) {
         this.repository = repository;
+        this.mapper = mapper;
     }
 
 
-    public DadosProcedimentoDTO salvar(CriarProcedimentoDTO dados) {
+    public DadosProcedimentoDTO salvarProcedimento(CriarProcedimentoDTO dados) {
         Procedimento procedimento = repository.save(mapper.toProcedimento(dados));
         return mapper.toDadosProcedimentoDTO(procedimento);
     }
@@ -35,18 +38,24 @@ public class ProcedimentoService {
         return mapper.toListDadosProcedimentoDTO(repository.findAll());
     }
 
-    public DadosProcedimentoDTO atualizar(String procedimentoId, AtualizarProcedimentoDTO update){
+    public DadosProcedimentoDTO mostrarProcedimento(String procedimentoId) {
+        return this.mapper.toDadosProcedimentoDTO(this.getProcedimentoById(procedimentoId));
+    }
+
+    public DadosProcedimentoDTO editarProcedimento(String procedimentoId, AtualizarProcedimentoDTO update){
         Procedimento alvo = this.getProcedimentoById(procedimentoId);
         alvo.atualizarDados(update.getNome(), update.getDescricao(), update.getValor());
         return mapper.toDadosProcedimentoDTO(repository.save(alvo));
     }
 
-    private Procedimento getProcedimentoById(String id) {
-        return this.repository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Procedimento não encontrado"));
+    public void deletarProcedimento(String id) {
+        if (repository.existsById(id))
+            throw new EntityNotFoundException(NOT_FOUND_MESSAGE);
+        repository.deleteById(id);
     }
 
-    public void remover(String id){
-        repository.deleteById(id);
+    public Procedimento getProcedimentoById(String id) {
+        return this.repository.findById(id)
+            .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_MESSAGE));
     }
 }
